Do load SHDefExtractIcon() during run-time for old MinGW
The problem with MinGW 5.3 is not just due to the missing function
declaration in the header, but also due to the function missing from
libshell32.a import library, so we do need to load it during run-time,
contrary to what the comment in the previous commit said.

This should finally fix MinGW build.