Fix expanding last wxDataViewCtrl column in the generic version
The code added in c0cf4939e59f47712bc22474d12f7a3f99e3fd4b didn't quite
work correctly because it used the old size of the window, before it was
resized, and not the new size. This was almost unnoticeable when
drag-resizing the window, but very noticeable when maximizing the
containing TLW, as could be seen on the 3rd page of the dataview sample,
for example, where the last column kept its size instead of expanding.

See #13904, #18295.