Preserve recent changes to wxSTC documentation
Recent commits 05163faabf and 41d5c958d3 fixed some typos is the
documentation for wxSTC. Unfortunatly these changes will be lost the
next time the wxSTC files are regenerated. This commit modifies
stc.interface.h.in and gen_docs.py to ensure that these changes are
preserved when the wxSTC files are regenerated.