Fix width returned by wxSpinCtrl::GetSizeFromTextSize() in wxMSW
Since the changes of e4e4df8e8c948362a0402173ee84fc08fe2143bb, this
function returned a value which was too small because the code in it
never accounted for the text control margins, but this was hidden by a
weird hack adding extra "3*y/10 + 10" pixels to the horizontal extent
which was removed by the commit above.

Really fix it now by reusing wxTextCtrl::GetSizeFromTextSize() which
correctly accounts for the margins and everything else.

See #12297.

Closes #18391.