Fix the issue with a last DVC column width on Mac
In Cocoa implementation of `wxDataViewControl`: call
`sizeLastColumnToFit` function of `NSTableView` in `OnSize` handler
(for any amount of columns, previously this function was called only
for a singe colunm DVCs).

`sizeLastColumnToFit` resizes the last column so the table view fits
exactly within its enclosing clip view.

This behavior is equivalent to `gtk` and `MSW` realizations.