Update wxImage used by Mac image graphics context in Flush()
Calling Flush() is supposed to make everything drawn so far immediately
visible, which means propagating it to the wxImage being drawn on in the
case of image-based graphics context, but wxMacCoreGraphicsImageContext
didn't do it, unlike Cairo and GDI+ version.

Fix this by overriding Flush() in it to explicitly update m_image.

Closes https://github.com/wxWidgets/wxWidgets/pull/1321